LangChain/Project1/requirement.txt
langchain_google_genai
langchain_core
streamlit
load_dotenv
python-dotenv
langchain_community